refactor: open-source readiness cleanup across barrel exports, imports, and module boundaries
Remove unconditional MLXBackend import from server barrel (crashes on non-Apple
platforms), enforce barrel-only imports in eval.py/runner.py, extract calibration
logic from eval_report.py (351→215 lines), add explicit temp=0.0 to MLX backend,
type-annotate runner dispatch, remove dead code branch and its tests, and update
core/ contract to reflect actual dependency policy.
