+++ tests/test_condense.py
@@ -14,7 +14,7 @@ from vaudeville.server.condense import (
     _split_into_chunks,
     condense_text,
 )
-from vaudeville.server import handle_request
+from vaudeville.server._handlers import handle_request
 
 
 class TestBuildCondensePrompt:
