+++ vaudeville/core/__init__.py
@@ -1,5 +1,11 @@
 from .client import VaudevilleClient
-from .protocol import ClassifyRequest, ClassifyResponse, parse_verdict
+from .protocol import (
+    ClassifyRequest,
+    ClassifyResponse,
+    ClassifyResult,
+    compute_confidence,
+    parse_verdict,
+)
 from .rules import (
     CHARS_PER_TOKEN,
     Rule,
@@ -15,8 +21,10 @@ __all__ = [
     "CHARS_PER_TOKEN",
     "ClassifyRequest",
     "ClassifyResponse",
+    "ClassifyResult",
     "Rule",
     "VaudevilleClient",
+    "compute_confidence",
     "load_rules",
     "load_rules_layered",
     "parse_rule",
