+++ vaudeville/server/__init__.py
@@ -1,10 +1,8 @@
 from .condense import condense_text
-from ._handlers import handle_request
 from .daemon import DaemonConfig, VaudevilleDaemon
 from .event_log import ClassificationEvent, EventLogger
 from .inference import InferenceBackend, LogprobBackend
 from .log_config import LogConfig, load_log_config
-from .mlx_backend import MLXBackend
 from .stats import aggregate_events, empty_result
 from .watch import watch
 
@@ -15,12 +13,10 @@ __all__ = [
     "InferenceBackend",
     "LogConfig",
     "LogprobBackend",
-    "MLXBackend",
     "VaudevilleDaemon",
     "aggregate_events",
     "condense_text",
     "empty_result",
-    "handle_request",
     "load_log_config",
     "watch",
 ]
